python/requirements.txt
click==6.7
Flask==0.12.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.10
SQLAlchemy==1.1.13
Werkzeug==0.12.2